Building blocks for packing finalised files.
This is the first in the series of commits that add support for packing
finalised files. The logic is pretty simple. On opening the DB, all the
finalised files are loaded into the in memory tree representation. When
the repack method is called, the records are dumped into a new packed
file, and the finalised files will be unlinked.
The dumping of records into the packed file and unlinking finalised
files is yet to be done, and will be part of the futur commits.