Use uint64_t instead of size_t.
`size_t` on a 32 bit machine is just 4 bytes, and assuming that we will
never run on a 32 bit machine was an oversight on my part. Fixing this.
I need to do a fair bit of testing to ensure that I haven't broken
anything.

There are still traces of `size_t`'s which I will replace in subsequent
commits.