Update the stored stat info.
When there's a change to the DB(like a new active file being created)
the `.zsdb` file is updated to reflect the change to the db. We need to
save the state so that we don't reload the DB unnecessarily.
This is a bug that I just discovered when running DB conversion under
callgrind.