DB files - always use calloc.
I had a `memset()` a static buffer going on for key/values less than
`65536` bytes. But on profiling it, with `callgrind`, I noticed that we
were hitting memset real hard and were wasting a fair number of CPU
cycles there. So, getting rid of it.

Now I basically will always use `calloc()` instead of `memset()`. In
most cases, they seem to be the same in the performance part.
Infact I have noticed that in some cases, `calloc()` seems to be doing less
work because there's no call to `memset()` (yes, `memset()` seems to be
that slow).

In comparison with `malloc()` + `memset()`, a call to `calloc()` is way
faster.