Remove calls to mfile_size.
A call to mfile_size() munmaps and mmaps the file, and this is done at
various parts in the code, which isn't desirable. This reflected in the
slowness in performing various operations on the DB.

The `size` field in a mfile(the memory mapped file) should reflect the
current size of the file, and we should use that to check the size.